Fix issue with pins falling through floor
Pins were falling through the floor when instantiated from the Renew function.  This was because the Lower function called afterwards pulled them through the floor because they were affected by gravity.  Turning off the pins' gravity on the prefab, then turning it back on once they were in place solved the issue.