Fix slight error in ScoreMaster
It was pointed out that the last line of the else if statement in the for loop was a perfromance waste because most of the time it would be adding zero.  I changed this line to an if statement and shuffled the responsibility of the iteration over the for loop to this if statement to keep everything concise.
I'm not sure that this code is really verbose enough, but the aim of ScoreMaster was to write the methods as concisely as possible, so this is the path that I went.  If I really wanted to save another line, I could move the edited line in this commit back into the iterator in the for loop declaration... but then it would just look silly!